rename Chip -> ServiceChip
