+++ web/src/app/alerts/CreateAlertDialog/StepContent/CreateAlertServiceSelect.tsx
@@ -20,7 +20,7 @@ import makeStyles from '@mui/styles/makeStyles'
 import ServiceLabelFilterContainer from '../../../services/ServiceFilterContainer'
 import { Search as SearchIcon } from '@mui/icons-material'
 import { FavoriteIcon } from '../../../util/SetFavoriteButton'
-import { ServiceChip } from '../../../util/Chips'
+import { ServiceChip } from '../../../util/ServiceChip'
 import AddIcon from '@mui/icons-material/Add'
 import _ from 'lodash'
 import getServiceFilters from '../../../util/getServiceFilters'
